requirements.txt
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
appdirs==1.4.4
async-timeout==5.0.1
attrs==25.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
cobra==0.29.1
colorama==0.4.6
depinfo==2.2.0
diskcache==5.6.3
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fair-esm==2.0.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gmpy2==2.1.5
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.6
importlib_resources==6.4.5
Jinja2==3.1.3
joblib==1.4.2
lightning-utilities==0.11.9
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
openpyxl==3.1.5
optlang==1.8.2
packaging==24.0
pandas==2.0.3
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
propcache==0.2.0
psutil==5.9.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==1.9.5
pytz==2025.1
PyYAML==6.0.1
rdkit==2024.3.5
regex==2024.11.6
requests==2.31.0
rich==13.9.4
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
sniffio==1.3.1
swiglpk==5.0.12
sympy==1.12
threadpoolctl==3.5.0
tokenizers==0.20.3
tqdm==4.66.2
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.1
yarl==1.15.2
